Created LiveSplit Client
- Created project
- Created LiveSplit Server Client
- Added barebones form flow for connecting to server and ending session
