+++ SaveGolds/SaveGolds.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31410.357
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SaveGolds", "SaveGolds\SaveGolds.csproj", "{A79226F7-B7E5-4D59-8171-A273BB1586E0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A79226F7-B7E5-4D59-8171-A273BB1586E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A79226F7-B7E5-4D59-8171-A273BB1586E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A79226F7-B7E5-4D59-8171-A273BB1586E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A79226F7-B7E5-4D59-8171-A273BB1586E0}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {EDD4B2E5-55CF-441F-A5CB-710D7EA28C2E}
+	EndGlobalSection
+EndGlobal
