+++ SaveGolds/SaveGolds/ConnectForm.cs
@@ -1,0 +1,50 @@
+﻿using System;
+using System.Windows.Forms;
+
+namespace SaveGolds
+{
+    public partial class ConnectForm : Form
+    {
+        public ConnectForm()
+        {
+            InitializeComponent();
+        }
+
+        private void connectButton_Click(object sender, EventArgs e)
+        {
+            string server = ipTextBox.Text;
+            int port = int.Parse(portTextBox.Text);
+            
+            try
+            {
+                var client = new LiveSplitClient(server, port);
+                var mainForm = new SaveGolds(client);
+                mainForm.Show();
+            } 
+            catch(InvalidOperationException)
+            {
+                ShowErrorBox(
+                    $"No viable IP Address found at {server}.",
+                    "No Viable IP"
+                );
+            }
+            catch(LiveSplitConnectionException connectionException)
+            {
+                ShowErrorBox(
+                    connectionException.Message,
+                    "Error Connecting to LiveSplit"
+                );
+            }
+        }
+
+        private void ShowErrorBox(string message, string caption)
+        {
+            MessageBox.Show(
+                message, 
+                caption, 
+                MessageBoxButtons.OK, 
+                MessageBoxIcon.Error
+            );
+        }
+    }
+}
